+++ docs/CHANGELOG.rst
@@ -4,10 +4,16 @@ CHANGELOG
 
 .. contents:: :local:
 
+
+1.5.0-alpha (in progress)
+=========================
+
+* csvchart, for making charts!
+
 1.4.0-alpha
 ===========
 
-* :doc:`csvpivot <moreutils/csvpivot>`_: Pivot a table on rows and columns (TK better label)
+:doc:`csvpivot <moreutils/csvpivot>`: Pivot a table on rows and columns (TK better label)
 
 
 1.3.7-alpha